requirements.txt
Django==1.6.1
Markdown==2.3.1
South==0.8.4
django-filter==0.7
django-guardian==1.1.1
django-tastypie==0.11.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.5.2
wsgiref==0.1.2
